+++ tests/util/test_calc.py
@@ -101,8 +101,8 @@ def test_run_ase_opt():
     atoms.calc = EMT()
 
     dyn = run_ase_opt(atoms, scratch_dir="test_calc", copy_files=["test_file.txt"])
-    Popen(f"gunzip {dyn.trajectory.filename}", shell=True).wait()
-    traj = read(zpath(dyn.trajectory.filename), index=":")
+    Popen(f"gunzip {dyn.trajectory.filename}.gz", shell=True).wait()
+    traj = read(dyn.trajectory.filename, index=":")
     assert traj[-1].calc.results is not None
     assert os.path.exists("test_file.txt")
     assert os.path.exists("test_file.txt.gz")
@@ -122,8 +122,8 @@ def test_run_ase_opt():
         copy_files=["test_file.txt"],
         optimizer_kwargs={"restart": None},
     )
-    Popen(f"gunzip {dyn.trajectory.filename}", shell=True).wait()
-    traj = read(zpath(dyn.trajectory.filename), index=":")
+    Popen(f"gunzip {dyn.trajectory.filename}.gz", shell=True).wait()
+    traj = read(dyn.trajectory.filename, index=":")
     assert traj[-1].calc.results is not None
 
     dyn = run_ase_opt(
@@ -134,8 +134,8 @@ def test_run_ase_opt():
         copy_files=["test_file.txt"],
         optimizer_kwargs={"restart": None, "trajectory": "new_test.traj"},
     )
-    Popen(f"gunzip {dyn.trajectory.filename}", shell=True).wait()
-    traj = read(zpath(dyn.trajectory.filename), index=":")
+    Popen(f"gunzip {dyn.trajectory.filename}.gz", shell=True).wait()
+    traj = read(dyn.trajectory.filename, index=":")
     assert traj[-1].calc.results is not None
 
     dyn = run_ase_opt(
@@ -149,8 +149,8 @@ def test_run_ase_opt():
             "trajectory": Trajectory("new_test2.traj", "w", atoms=traj[-1]),
         },
     )
-    Popen(f"gunzip {dyn.trajectory.filename}", shell=True).wait()
-    traj = read(zpath(dyn.trajectory.filename), index=":")
+    Popen(f"gunzip {dyn.trajectory.filename}.gz", shell=True).wait()
+    traj = read(dyn.trajectory.filename, index=":")
     assert traj[-1].calc.results is not None
 
     with pytest.raises(ValueError):