+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added a CLI for `quacc` along with a `quacc config` option to configure Covalent appropriately upon install.
 - Added a `quacc.util.wflows.make_dask_cluster` function to easily generate a Dask cluster.
 - Added generic type hints for schemas.
+- Added a `MAKE_UNIQUE_WORKDIR` global setting to have quacc automatically make a unique working directory for each calculation.
 
 ### Changed
 
@@ -29,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - Fixed a bug where parallel, multithreaded Python jobs could cause file I/O issues.
+- Temporary directories are cleaned up after the run is completed.
 
 ### Removed